requirements.txt
Flask==0.12.2
Pillow==5.0.0
celery==4.1.0
redis==2.10.6
weasyprint==0.42
xmltodict==0.11.0
